remove legacy unused code
